Add M4 channel mapping version number
To recognize if the sent mapping is compatible with the autopilot version I send a fixed bit pattern in the unused 4 bits of channel 9.
Every channel has 12 bits for it's value. Thanks to the raw mapping the exact, unaltered values are delivered to the autopilot.
On channel 9 the first 5 bits are in use for the 2-way switch and buttons. I used the last 4 bits (most significant) for the version tag.
I just add a mapping of the first 3-way switch and no matter in what position it is 0x0F00 gets added to the channel.
From that we can read the first "version tag" is 0xF which can be decreased for other versions in the future.
I chose 0xF because the probability of this value staying on these 4 bits with a legacy mapping is very low
and hence it's a good check for compatibility.